requirements.txt
aiohttp==3.9.3
aiosignal==1.3.1
annotated-types==0.6.0
anyio==4.3.0
appdirs==1.4.4
async-timeout==4.0.3
attrs==22.1.0
certifi==2022.6.15
charset-normalizer==2.1.0
click==8.1.3
coverage==6.2
cycler==0.11.0
docker-pycreds==0.4.0
exceptiongroup==1.2.0
fastapi==0.110.0
filelock==3.13.1
flake8==4.0.1
frozenlist==1.4.1
fsspec==2024.2.0
gitdb==4.0.11
GitPython==3.1.42
h11==0.14.0
idna==3.3
imageio==2.21.1
iniconfig==1.1.1
Jinja2==3.1.3
joblib==1.1.0
kiwisolver==1.4.4
lightning==2.2.1
lightning-utilities==0.10.1
MarkupSafe==2.1.5
matplotlib
mccabe==0.6.1
mpmath==1.3.0
multidict==6.0.5
networkx==2.8.5
numpy
nvidia-cublas-cu12==12.1.3.1
nvidia-cuda-cupti-cu12==12.1.105
nvidia-cuda-nvrtc-cu12==12.1.105
nvidia-cuda-runtime-cu12==12.1.105
nvidia-cudnn-cu12==8.9.2.26
nvidia-cufft-cu12==11.0.2.54
nvidia-curand-cu12==10.3.2.106
nvidia-cusolver-cu12==11.4.5.107
nvidia-cusparse-cu12==12.1.0.106
nvidia-nccl-cu12==2.19.3
nvidia-nvjitlink-cu12==12.3.101
nvidia-nvtx-cu12==12.1.105
opencv-python==4.5.5.62
packaging==21.3
pandas==1.4.3
patsy==0.5.2
Pillow==9.2.0
plotly==5.10.0
plotly-express==0.4.1
pluggy==1.0.0
plyfile==0.7.4
prettytable==3.9.0
protobuf==3.19.4
psutil==5.9.8
py==1.11.0
pycodestyle==2.8.0
pydantic==2.6.3
pydantic-core==2.16.3
pyflakes==2.4.0
pyparsing==3.0.9
pytest==6.2.5
pytest-cov==3.0.0
pytest-lazy-fixture==0.6.3
python-dateutil==2.8.2
pytorch-lightning==2.2.1
pytz==2022.2.1
PyWavelets==1.3.0
PyYAML==6.0.1
requests==2.28.1
scikit-image==0.19.1
scikit-learn
scipy
seaborn==0.11.2
sentry-sdk==1.40.6
setproctitle==1.3.3
six==1.16.0
smmap==5.0.1
sniffio==1.3.1
starlette==0.36.3
statsmodels==0.13.2
sympy==1.12
tenacity==8.0.1
termcolor==1.1.0
threadpoolctl==3.1.0
tifffile==2022.8.12
toml==0.10.2
torchmetrics==1.3.1
tqdm==4.62.3
triton==2.2.0
typing-extensions==4.10.0
urllib3==1.26.11
uvicorn==0.27.1
wandb==0.16.4
wcwidth==0.2.12
xgboost==2.0.3
xmltodict==0.13.0
yarl==1.9.4
